+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Dialog Insight: Make birthdate value optional when sending to web service
+* Markup "modifyUrl" now goes to cancel page, not contact page
 
 ### 1.4.3: May 24th, 2021
 * Times dropdown must be above header in fixed position